Remove if collides from Code API
